requirements.txt
line-bot-sdk
Flask==1.1.2
flask_cors==3.0.10
flask_uploads==0.2.1
gunicorn==19.10.0
requests-oauthlib==1.3.0
requests2==2.16.0
requests >= 2.0
future